requirements.txt
pandas==2.3.2
requests==2.32.5

strands-agents>=1.0.0
strands-agents-tools>=0.2.0

python-dotenv==1.1.1
